changing the class to Helper and moving methods from MainForm class to Helper class
